systems/Lotus/requirements.txt
# Requires-Python >=3.10,<3.13

pandas>=2.0.0
numpy>=1.24.0
sql_metadata

lotus-ai

litellm
litellm[google]
pydantic
tqdm
tokenizers

python-dotenv